# backend/docs/swagger.yaml
basePath: /
definitions:
...
  web-hosting_internal_modules_kurikulum_dto.JurusanResponse:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: teknik-elektro
        type: string
    type: object
  web-hosting_internal_modules_kurikulum_dto.KurikulumCreateRequest:
    properties:
      kode:
        example: myhutao-2024
        maxLength: 12
        type: string
      name:
        example: Merdeka Belajar
        maxLength: 255
        type: string
      prodi_name:
        example: teknik-informatika
        maxLength: 255
        type: string
    required:
    - kode
    - name
    - prodi_name
    type: object
  web-hosting_internal_modules_kurikulum_dto.KurikulumMKResponse:
    properties:
      mata_kuliah:
        $ref: '#/definitions/web-hosting_internal_modules_kurikulum_dto.MataKuliahResponse'
      semester:
        example: 3
        type: integer
      wajib:
        example: true
        type: boolean
    type: object
  web-hosting_internal_modules_kurikulum_dto.KurikulumResponse:
    properties:
      id:
        example: 0000-0000
        type: string
      kode:
        example: my-hutao
        type: string
      kurikulum_mk:
        items:
          $ref: '#/definitions/web-hosting_internal_modules_kurikulum_dto.KurikulumMKResponse'
        type: array
      name:
        example: My Hutao
        type: string
      prodi:
        $ref: '#/definitions/web-hosting_internal_modules_kurikulum_dto.ProdiResponse'
    type: object
  web-hosting_internal_modules_kurikulum_dto.KurikulumUpdateRequest:
    properties:
      kode:
        example: myhutao-2024
        maxLength: 12
        type: string
      name:
        example: Merdeka Belajar
        maxLength: 255
        type: string
      prodi_name:
        example: teknik-informatika
        maxLength: 255
        type: string
    type: object
  web-hosting_internal_modules_kurikulum_dto.MataKuliahResponse:
    properties:
      id:
        example: MK001
        type: string
      kode:
        example: MK001
        type: string
      name:
        example: Pemrograman Berorientasi Objek
        type: string
      sks:
        example: 3
        type: integer
    type: object
...
  web-hosting_internal_modules_kurikulum_dto.ProdiResponse:
    properties:
      id:
        example: 1
        type: integer
      jenjang:
        example: D3
        type: string
      jurusan:
        $ref: '#/definitions/web-hosting_internal_modules_kurikulum_dto.JurusanResponse'
      name:
        example: teknik informatika
        type: string
    type: object
...
  web-hosting_internal_package_swagger.ErrCreateKurikulumFailed:
    properties:
      error:
        example: kurikulum already exists
        type: string
      message:
        example: failed to create kurikulum
        type: string
      path:
        example: /api/kurikulum
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrCreateKurikulumInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to create kurikulum
        type: string
      path:
        example: /api/kurikulum
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrDeleteKurikulumFailed:
    properties:
      error:
        example: kurikulum already exists
        type: string
      message:
        example: failed to delete kurikulum
        type: string
      path:
        example: /api/kurikulum
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrDeleteKurikulumInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to delete kurikulum
        type: string
      path:
        example: /api/kurikulum
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrForbiddenAccess:
    properties:
      error:
        example: Forbidden
        type: string
      message:
        example: Role anda tidak diizinkan
        type: string
      path:
        example: /api/super/role
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrGetKurikulumFailed:
    properties:
      error:
        example: kurikulum not found
        type: string
      message:
        example: failed to get kurikulum
        type: string
      path:
        example: /api/kurikulum
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrGetKurikulumInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to get kurikulum
        type: string
      path:
        example: /api/kurikulum
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken:
    properties:
      error:
        example: invalid token
        type: string
      message:
        example: failed_auth
        type: string
      path:
        example: /api/auth/logout
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrUpdateKurikulumFailed:
    properties:
      error:
        example: kurikulum already exists
        type: string
      message:
        example: failed to update kurikulum
        type: string
      path:
        example: /api/kurikulum
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrUpdateKurikulumInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to update kurikulum
        type: string
      path:
        example: /api/kurikulum
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_utils.PaginatedData-array_web-hosting_internal_modules_kurikulum_dto_KurikulumResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/web-hosting_internal_modules_kurikulum_dto.KurikulumResponse'
        type: array
      pagination:
        $ref: '#/definitions/web-hosting_internal_package_utils.PaginationMeta'
    type: object
...
  web-hosting_internal_package_utils.PaginationMeta:
    properties:
      page:
        example: 1
        type: integer
      per_page:
        example: 10
        type: integer
      total_items:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  web-hosting_internal_package_utils.Response-any-any:
    properties:
      data: {}
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
  web-hosting_internal_package_utils.Response-web-hosting_internal_modules_kurikulum_dto_KurikulumResponse-any:
    properties:
      data:
        $ref: '#/definitions/web-hosting_internal_modules_kurikulum_dto.KurikulumResponse'
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
  ? web-hosting_internal_package_utils.Response-web-hosting_internal_package_utils_PaginatedData-array_web-hosting_internal_modules_kurikulum_dto_KurikulumResponse-any
  : properties:
      data:
        $ref: '#/definitions/web-hosting_internal_package_utils.PaginatedData-array_web-hosting_internal_modules_kurikulum_dto_KurikulumResponse'
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    **BARANG SIAPA YANG UPDATE/HAPUS DATA SEMBARANGAN DAPAT KECUPAN DARI HAFIZ**.
    **BUAT DATA BARU JIKA INGIN HAPUS/UPDATE**.

    **DUMMY DATA LINK:**
    **https://github.com/Kar-Su/pbl-project-simpadu/tree/main/backend/internal/database/seeders/json**

    **TUTORIAL LINK:***
    **https://github.com/Kar-Su/pbl-project-simpadu**
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TIM 1 API
  version: "1.0"
paths:
...
  /api/kurikulum:
    get:
      consumes:
      - application/json
      description: |-
        Mendapatkan data kurikulum. Jika path kode diberikan, return 1 data; tanpa kode, return semua dengan pagination

        **Akses:** Logged User
      parameters:
      - description: 'Halaman (default: 1, per halaman: 10)'
        example: 1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-web-hosting_internal_package_utils_PaginatedData-array_web-hosting_internal_modules_kurikulum_dto_KurikulumResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetKurikulumFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetKurikulumInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Get Kurikulum
      tags:
      - kurikulum
    post:
      consumes:
      - application/json
      description: |-
        Menambahkan kurikulum baru ke sistem

        **Akses:** Admin Akademik
      parameters:
      - description: Kurikulum Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web-hosting_internal_modules_kurikulum_dto.KurikulumCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-any-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrCreateKurikulumFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrCreateKurikulumInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Create Kurikulum Baru
      tags:
      - kurikulum
  /api/kurikulum/{kode}:
    delete:
      consumes:
      - application/json
      description: |-
        Menghapus kurikulum berdasarkan kode di path parameter

        **Akses:** Admin Akademik
      parameters:
      - description: Kode Kurikulum
        example: myhutao-2024
        in: path
        name: kode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-any-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrDeleteKurikulumFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrDeleteKurikulumInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Delete Kurikulum berdasarkan kode
      tags:
      - kurikulum
    get:
      consumes:
      - application/json
      description: |-
        Mendapatkan data kurikulum. Jika path kode diberikan, return 1 data; tanpa kode, return semua dengan pagination

        **Akses:** Logged User
      parameters:
      - description: Kode Kurikulum (opsional)
        example: myhutao-2024
        in: path
        name: kode
        type: string
      - description: 'Halaman (default: 1, per halaman: 10)'
        example: 1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-web-hosting_internal_package_utils_PaginatedData-array_web-hosting_internal_modules_kurikulum_dto_KurikulumResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetKurikulumFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetKurikulumInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Get Kurikulum
      tags:
      - kurikulum
    put:
      consumes:
      - application/json
      description: |-
        Mengupdate kurikulum berdasarkan kode di path parameter

        **Akses:** Admin Akademik
      parameters:
      - description: Kode Kurikulum
        example: myhutao-2024
        in: path
        name: kode
        required: true
        type: string
      - description: Kurikulum Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web-hosting_internal_modules_kurikulum_dto.KurikulumUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-web-hosting_internal_modules_kurikulum_dto_KurikulumResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUpdateKurikulumFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUpdateKurikulumInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Update Kurikulum berdasarkan kode
      tags:
      - kurikulum
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Format: Bearer <access token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"